Extend completions model
Instead of assuming that all completions are of wordlist type (i.e.,
invoke compgen -W), create a structure to hold completions that can
account for wordlists or other flags to be passed to compgen.

Note that this commit implements changes to the compgen scripts for Bash
and ZSH. Documentation suggests that this step be automated in a build
pipeline for installation by package, or installed in a user's profile
by evaulating the script directly. In either of these cases, the upgrade
path should be smooth.