Handle --version just like --help
when the version flag is passed, the application should do nothing but
print the version and exit. But currently, the pre-action for the
version flag got executed AFTER the default for all the other flags were
set.

Setting defaults involves validation like checking for file's
existence etc. These validation are not necessary when the --version
flag is passed. This commit fixes that